+++ src/main/java/com/nohit/jira_project/model/GioHang.java
@@ -14,7 +14,6 @@ import static javax.persistence.FetchType.*;
 @NoArgsConstructor
 public class GioHang {
     @Id
-    @GeneratedValue
     @Column(name = "id")
     private int id;
